Nucleotide count - frequency of characters in a string
Count how many of each nucleotide is in a DNA sequence

The sequence is represented as a Clojure string.  Each character is a
nucleotide.

Show different approaches to solving the problem at various levels of
abstraction.

Use `frequencies` function to generate a map of how often each
nucleotide (character) appears in the DNA sequence (string).