Bank Account - parrallelism with a Clojure atom
Simulate a bank account supporting opening/closing, withdrawals, and deposits of
money, which supports concurrent transactions.

As functions will be called concurrently, a Clojure atom is used to manage any
state changes of the bank accounts.  This ensures that the bank account can only
be changed by one call at any one time.

The bank account was defined as a hash-map, with user-id as a UUID (universal
id) from the java.util library.