Rna Transcription: revised solution, using or instead of if
replaced apply str with clojure.string/join

replaced parent if statement that only checks for "X" as bad nucleotie. Use a
hash-map as conditional with the `or` function, to act as a check for correct
nucleotides and convert them with the same expression.

If the given dna nucleotide is not in the hash-map, then the AssertionError is
called.