RNATranscription: pure function and supporting tests
Add transcription map as an argument to the dna->rna transforming function,
making that function pure and able to accept alternative transformations.

Add a new namespace for the pure dna->rna function, then a matching test
namespace can be created.  This avoids having to change the existing test
namespace that we received from Exercism